refactor(query_planning): colocate outcome classification with the enum
Move the QueryPlannerError -> QueryPlanningOutcome conversion next to the enum
in query_planner_service.rs, and add a From<&FederationErrorBridge> impl that
the QueryPlannerError conversion delegates to. plan.duration now classifies via
that shared impl and makes a single metric_query_planning_plan_duration call.
Document memory_limit as part of the shared outcome vocabulary.
