+++ docs/source/routing/observability/router-telemetry-otel/enabling-telemetry/standard-instruments.mdx
@@ -181,7 +181,7 @@ The `apollo.router.cache.redis.errors` metric also includes an `error_type` attr
 
 - `apollo.router.query_planning.warmup.duration` - Time spent warming up the query planner queries in seconds.
 - `apollo.router.query_planning.warmup.operations` - Counter of operations processed during query planner warm-up. Together with `apollo.router.query_planning.warmup.operations.expected`, this lets you compute warm-up coverage (planned / expected). Attributes:
-  - `outcome`: The result of planning the operation. Uses the same values as `apollo.router.query_planning.plan.duration` (`success`, `timeout`, `cancelled`, `error`, `memory_limit`); warm-up currently emits `success`, `timeout`, and `error`.
+  - `outcome`: The result of planning the operation. Uses the shared query-planning outcome values (`success`, `timeout`, `cancelled`, `error`, `memory_limit`), the same set as `apollo.router.query_planning.plan.duration`.
   - `source`: Where the operation came from (`persisted_query` for the persisted-query manifest, `cache` for operations carried over from the previous in-memory cache)
 - `apollo.router.query_planning.warmup.operations.expected` - Counter of operations that warm-up intended to plan. Attributes:
   - `source`: Where the operation came from (`persisted_query`, `cache`)
@@ -190,7 +190,7 @@ The `apollo.router.cache.redis.errors` metric also includes an `error_type` attr
   - `phase`: The warm-up phase where backpressure occurred (`parse` for document parsing, `plan` for the planner call)
 - `apollo.router.query_planning.plan.duration` - Histogram of plan durations isolated to query planning time only. Attributes:
   - `planner`: The query planner implementation that produced the plan (for example, `rust`)
-  - `outcome`: The result of planning the operation (`success`, `timeout`, `cancelled`, `error`)
+  - `outcome`: The result of planning the operation (`success`, `timeout`, `cancelled`, `error`, `memory_limit`)
 - `apollo.router.query_planning.total.duration` - Histogram of plan durations including queue time.
 - `apollo.router.query_planning.plan.evaluated_plans` - Histogram of the number of evaluated query plans.
 - `apollo.router.query_planning.plan.evaluated_paths` - Histogram of the number of paths (including intermediate ones) the planner considers before generating a plan. High values often correlate with long planning times on complex schemas or queries. Tune the limits as described in [Tuning query planner limits](/graphos/routing/query-planning/query-planning-best-practices#tuning-query-planner-limits).
